(PC-6910): Move offerer model to core.offerer
