+++ src/pcapi/core/users/api.py
@@ -15,6 +15,7 @@ from jwt import InvalidTokenError
 from pcapi import settings
 from pcapi.core import mails
 from pcapi.core.bookings.conf import LIMIT_CONFIGURATIONS
+from pcapi.core.offerers.offerer import Offerer
 from pcapi.core.payments import api as payment_api
 from pcapi.core.users.models import Credit
 from pcapi.core.users.models import DomainsCredit
@@ -40,7 +41,6 @@ from pcapi.emails.beneficiary_email_change import build_beneficiary_information_
 from pcapi.models import BeneficiaryImport
 from pcapi.models import ImportStatus
 from pcapi.models.db import db
-from pcapi.models.offerer import Offerer
 from pcapi.models.user_offerer import UserOfferer
 from pcapi.models.user_session import UserSession
 from pcapi.repository import repository
