+++ src/pcapi/sandboxes/scripts/getters/pro_12_bookings.py
@@ -1,9 +1,9 @@
 from pcapi.core.bookings.models import Booking
+from pcapi.core.offerers.offerer import Offerer
 from pcapi.core.offers.models import Offer
 from pcapi.core.offers.models import Stock
 from pcapi.core.users.models import User
 from pcapi.models import Venue
-from pcapi.models.offerer import Offerer
 from pcapi.models.user_offerer import UserOfferer
 from pcapi.sandboxes.scripts.utils.helpers import get_booking_helper
 from pcapi.sandboxes.scripts.utils.helpers import get_offer_helper
