+++ src/pcapi/sandboxes/scripts/getters/pro_13_reimbursements.py
@@ -1,9 +1,9 @@
 from pcapi.core.bookings.models import Booking
+from pcapi.core.offerers.offerer import Offerer
 from pcapi.core.offers.models import Offer
 from pcapi.core.offers.models import Stock
 from pcapi.core.users.models import User
 from pcapi.models import Venue
-from pcapi.models.offerer import Offerer
 from pcapi.models.payment import Payment
 from pcapi.models.user_offerer import UserOfferer
 from pcapi.sandboxes.scripts.utils.helpers import get_booking_helper
